+++ tests/testthat/test-read-parquet-2.R
@@ -141,6 +141,15 @@ test_that("FLOAT", {
   })
 })
 
+test_that("FLOAT from multiple row groups and pages", {
+  skip_on_cran()
+  pf <- test_path("data/float.parquet")
+  expect_equal(
+    as.data.frame(arrow::read_parquet(pf)),
+    as.data.frame(read_parquet(pf))
+  )
+})
+
 test_that("DOUBLE", {
   tmp <- tempfile(fileext = ".parquet")
   on.exit(unlink(tmp), add = TRUE)