+++ NEWS.md
@@ -3,6 +3,9 @@
 * `write_parquet()` now correctly converts double `Date` columns
   to integer columns (@eitsupi, #116).
 
+* `read_parquet()` now correctly reads `FLOAT` columns from files with
+  multiple row groups.
+
 # nanoparquet 0.4.0
 
 * API changes: